+++ src/db/writebatch.h
@@ -20,7 +20,7 @@ class WriteBatch {
   void Clear();
   void Append(WriteBatch* source);
   // State Iterate(Handler* handler) const;
-  size_t mateSize() const { return mate.size(); }
+  uint32_t mateSize() const { return mate.size(); }
   // 返回条目数
   int Count();
   // 修改条目数
@@ -30,7 +30,7 @@ class WriteBatch {
   void SetSequence(SequenceNum seq);
 
   std::string_view Contents() { return std::string_view(mate); }
-  size_t ByteSize() { return mate.size(); }
+  uint32_t ByteSize() { return mate.size(); }
 
   void SetContents(const std::string_view& contents);
